api/requirements.txt
fastapi[standard]
transformers
torch
sqlalchemy
psycopg2-binary
python-dotenv